Renaming DangerousMethodUsageAnalyzer to DangerousMemberUsagesAnalyzer in preparation for methods
