+++ tests/D2L.CodeStyle.Analyzers.Test/Specs/DangerousMemberUsagesAnalyzer.DangerousMethods.cs
@@ -1,4 +1,4 @@
-﻿// analyzer: D2L.CodeStyle.Analyzers.ApiUsage.DangerousMethodUsages.DangerousMethodUsagesAnalyzer
+﻿// analyzer: D2L.CodeStyle.Analyzers.ApiUsage.DangerousMemberUsages.DangerousMemberUsagesAnalyzer
 
 using System;
 using System.Reflection;
